middlewares error fix

+++ backend/utils/errorHandler.js
@@ -1,18 +1,22 @@
+const ApiError = require("./createApiError");
+const ApiResponse = require("./ApiResponse");
 const fs = require("fs");
 const errorHandler = (error, req, res, next) => {
     console.log("=========================");
     console.log(error.stack);
     console.log("=========================");
-    const statusCode = error.statusCode || 500;
+    if (error instanceof ApiError) {
+        return res
+            .status(error.statusCode)
+            .json(new ApiResponse(error.statusCode, error.message));
+    }
+    const statusCode = 500;
     if (req.file && req.file.path) {
         fs.unlink(req.file.path, (err) => {
             if (err) console.error("Failed to delete file:", err);
         });
     }
-    res.status(statusCode).json({
-        status: false,
-        message: error.message,
-    });
+    res.status(statusCode).json(new ApiResponse(statusCode, error.message));
 };
 
 module.exports = errorHandler;
